requirements.txt
pyinstaller>=5.0.0
python-dotenv>=1.0.0
prettytable>=3.0.0
